Create donateETH.sol
This is a smart contract in Solidity which accepts donations in ETH, and allows the contract owner to withdraw the funds. Be sure to save the PriceConverter.sol library into your Remix IDE before deploying so that your donateETH contract is able to accept a minimum donation of 5 USD and above.