Create PriceConverterLibrary
This PriceConverter library uses Chainlink decentralized data feeds to fetch the latest price of ETH/USD from AggregatorV3Interface. This library is used in the donateETH.sol contract to fetch the latest price of ETH, and requires the sender to donate a minimum of 5USD in ETH at its current price.